+++ lib/i18n.d.ts
@@ -1,12 +1,9 @@
-import React from 'react';
 import { Translations, ConvertMissingKey, Postprocess, Transforms, Preprocess } from './';
-interface I18nContext {
-    locale: string;
-    globalTranslations?: Translations;
+export interface I18nContext<L extends string = string> {
+    locale: L;
+    globalTranslations?: Translations<L>;
     convertMissingKey?: ConvertMissingKey;
     transforms?: Transforms;
     preprocess?: Preprocess;
     postprocess?: Postprocess;
 }
-export declare const I18n: React.Context<I18nContext>;
-export {};