+++ lib/index.d.ts
@@ -1,31 +1,32 @@
-/// <reference types="react" />
-export { I18n } from './i18n';
+export { I18nContext } from './i18n';
 export { default as translate } from './hoc';
 export * from './hooks';
 export { default as tr } from './translate';
-export declare type Translations = Record<string, {
-    [language: string]: string;
+export declare type Translations<L extends string = string> = Record<string, {
+    [language in L]: string;
 }>;
 export declare type ConvertMissingKey = (key: string, translations: Translations) => string;
 export declare type Preprocess = (key: string, translations: Translations) => string | boolean | undefined | null | void;
-export declare type Postprocess = (phrase: string) => string | React.ReactNode;
+export declare type Postprocess = (phrase: string) => string;
 export interface Transforms {
     [name: string]: (value: any) => string;
 }
-export interface TranslateProps {
-    t: TranslateLocal;
-    g: TranslateGlobal;
-    getLocale(): string;
+export interface TranslateProps<L extends string = string, T extends Translations<L> = Translations<L>> {
+    t: TranslateLocal<L, T>;
+    g: TranslateGlobal<L, T>;
+    getLocale(): L;
 }
-export interface TranslateType<T extends Translations = Translations> {
-    (key: keyof T, interpolation?: Interpolation): string;
-    locale: string;
+export interface TranslateType<L extends string = string, T extends Translations<L> = Translations<L>> {
+    (key: keyof T & string, interpolation?: Interpolation): string;
+    locale: L;
     has(key: string): boolean;
 }
-export interface TranslateLocal<T extends Translations = Translations> extends TranslateType<T> {
-    g: TranslateType<T>;
+export interface TranslateLocal<L extends string = string, T extends Translations<L> = Translations<L>> extends TranslateType<L, T> {
+    g: TranslateType<L, T>;
 }
-export declare type TranslateGlobal = TranslateType;
-export declare type Interpolation = {
-    [key: string]: string | number | object;
-} | number;
+export declare type TranslateGlobal<L extends string = string, T extends Translations<L> = Translations<L>> = TranslateType<L, T>;
+export declare type Interpolation = ({
+    smart_count?: number;
+} & {
+    [key: string]: string | object;
+}) | number;