+++ lib/translate.d.ts
@@ -1,10 +1,10 @@
 import { Translations, ConvertMissingKey, Transforms, TranslateLocal, Postprocess, Preprocess } from '.';
-export default function translate<T extends Translations>({ locale, translations, globalTranslations, convertMissingKey, transforms, preprocess, postprocess, }: {
-    locale: string;
+export default function translate<L extends string = string, T extends Translations = Translations<L>>({ locale, translations, globalTranslations, convertMissingKey, transforms, preprocess, postprocess, }: {
+    locale: L;
     translations: T;
-    globalTranslations?: Translations;
+    globalTranslations?: Translations<L>;
     convertMissingKey?: ConvertMissingKey;
     transforms?: Transforms;
     preprocess?: Preprocess;
     postprocess?: Postprocess;
-}): TranslateLocal<T>;
+}): TranslateLocal<L, T>;