+++ src/hoc.tsx
@@ -1,26 +1,32 @@
 import React, { useMemo, useContext } from 'react';
-import { I18n, Translations, TranslateProps } from './';
+import { Translations, TranslateProps } from './';
+import { I18nContext } from './i18n';
 import translate from './translate';
 
-const hoc = <T extends Translations>(translations: T) => <Props extends object>(
-  WrappedComponent: React.ComponentType<Props & TranslateProps>
-) => {
-  const LocalTranslations: React.FC<Props> = props => {
-    const ctx = useContext(I18n);
-    const i18nProps = useMemo(() => {
-      const t = translate({ ...ctx, translations });
+const hoc =
+  <L extends string, T extends Translations<L>>(
+    Context: React.Context<I18nContext<L>>,
+    translations: T
+  ) =>
+  <Props extends object>(
+    WrappedComponent: React.ComponentType<Props & TranslateProps<L, T>>
+  ) => {
+    const LocalTranslations: React.FC<Props> = (props) => {
+      const ctx = useContext(Context);
+      const i18nProps = useMemo(() => {
+        const t = translate({ ...ctx, translations });
 
-      return {
-        t: t,
-        g: t.g,
-        getLocale: () => ctx.locale,
-      };
-    }, [ctx]); // translations is always constant
+        return {
+          t: t,
+          g: t.g,
+          getLocale: () => ctx.locale,
+        };
+      }, [ctx]); // translations is always constant
 
-    return <WrappedComponent {...props} {...i18nProps} />;
+      return <WrappedComponent {...props} {...i18nProps} />;
+    };
+
+    return LocalTranslations;
   };
-
-  return LocalTranslations;
-};
 
 export default hoc;