+++ src/hooks.ts
@@ -1,14 +1,18 @@
 import { useContext, useMemo } from 'react';
-import { I18n } from './i18n';
 import translate from './translate';
-import { Translations } from '.';
+import { I18nContext, Translations } from '.';
 
-export function useTranslations<T extends Translations>(translations: T) {
-  const ctx = useContext(I18n);
-  return useMemo(() => translate({ ...ctx, translations }), [ctx, translations]);
+export function useTranslations<
+  L extends string = string,
+  T extends Translations<L> = Translations<L>
+>(Context: React.Context<I18nContext<L>>, translations: T) {
+  const ctx = useContext(Context);
+  return useMemo(() => translate<L, T>({ ...ctx, translations }), [ctx, translations]);
 }
 
-export function useLocale(): string {
-  const { locale } = useContext(I18n);
+export function useLocale<L extends string = string>(
+  Context: React.Context<I18nContext<L>>
+) {
+  const { locale } = useContext(Context);
   return locale;
 }